Handing in,
Zombie spawn generation is not complete, so zombies will not spawn once the initial ten are killed.

Bullet velocities are also not complete, whenever a bullet is shot, it changes the velocities of all bullets on the screen, because of this the game sometimes crashes when bullets intersect in a certain way.

Other then that the concept is simple, run around and kill zombies...

WASD to move, player aims at mouse cursor, left mouse button to shoot.

I don't know why the player is Megaman

2017 / 06 / 20

- Maloof